refactor(api): move getApp to server/api/test/utils
